simp nft algo config
