+++ tests/special_e2e/run_diffusionnft_qwen_image.sh
@@ -75,10 +75,10 @@ python3 -m verl_omni.trainer.main_diffusion \
     actor_rollout_ref.ref.log_prob_micro_batch_size_per_gpu=${micro_bsz_per_gpu} \
     algorithm.trainer_type=direct_preference \
     algorithm.sample_source=online \
-    algorithm.diffusion_nft.timestep_fraction=1.0 \
-    algorithm.diffusion_nft.old_policy_decay_type=1 \
-    algorithm.diffusion_nft.old_policy_update_interval=1 \
-    algorithm.diffusion_nft.adv_mode=continuous \
+    algorithm.algo_config.timestep_fraction=1.0 \
+    algorithm.algo_config.old_policy_decay_type=1 \
+    algorithm.algo_config.old_policy_update_interval=1 \
+    algorithm.algo_config.adv_mode=continuous \
     reward.num_workers=1 \
     reward.reward_model.enable=False \
     trainer.logger=console \
