+++ verl_omni/trainer/config/algorithm.py
@@ -15,35 +15,10 @@
 """Diffusion-specific algorithm config additions for verl_omni."""
 
 from dataclasses import dataclass, field
-from typing import Optional
 
 from verl.base_config import BaseConfig
 
-__all__ = ["DiffusionNFTAlgoConfig", "DiffusionAlgoConfig"]
-
-
-@dataclass
-class DiffusionNFTAlgoConfig(BaseConfig):
-    """DiffusionNFT-specific algorithm controls."""
-
-    old_policy_decay_type: int = 0
-    old_policy_decay: Optional[float] = None
-    old_policy_update_interval: int = 1
-    timestep_fraction: float = 1.0
-    adv_mode: str = "continuous"
-
-    def __post_init__(self):
-        valid_adv_modes = {"continuous", "positive_only", "negative_only", "one_only", "binary"}
-        if self.adv_mode not in valid_adv_modes:
-            raise ValueError(
-                f"Invalid DiffusionNFT adv_mode: {self.adv_mode}. Must be one of {sorted(valid_adv_modes)}"
-            )
-        if self.old_policy_update_interval <= 0:
-            raise ValueError(
-                f"DiffusionNFT old_policy_update_interval must be positive, got {self.old_policy_update_interval}."
-            )
-        if not 0 < self.timestep_fraction <= 1:
-            raise ValueError(f"DiffusionNFT timestep_fraction must be in (0, 1], got {self.timestep_fraction}.")
+__all__ = ["DiffusionAlgoConfig"]
 
 
 @dataclass
@@ -56,4 +31,4 @@ class DiffusionAlgoConfig(BaseConfig):
     norm_adv_by_std_in_grpo: bool = True
     bypass_mode: bool = False
     global_std: bool = True
-    diffusion_nft: DiffusionNFTAlgoConfig = field(default_factory=DiffusionNFTAlgoConfig)
+    algo_config: BaseConfig = field(default_factory=BaseConfig)
